% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustMultipleTesting.R
\name{adjustMultipleTesting}
\alias{adjustMultipleTesting}
\title{Calculate adjusted pvalues}
\usage{
adjustMultipleTesting(pval_list, alpha)
}
\arguments{
\item{pval_list}{Pvalues from multiple testing}

\item{alpha}{threshold}
}
\value{
A list object, which contains the following two items:
\itemize{
\item \bold{\code{Table}} A data frame object with columns of rank, pvals, adjusted_pvals and whether to reject null hypothesis.
\item \bold{\code{Proportion}} The proportion of the tests that reject the null hypothesis H0 among all the tests.
}
}
\description{
Calculate adjusted pvalues of multiple testing.
}
\examples{
file <- system.file(package="RgnTX", "extdata/multi_pvals.rds")
multi_pvals <- readRDS(file)
adjustMultipleTesting(multi_pvals[, 1], 0.05)
}
