% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShiftedZScoreTx.R
\name{plotShiftedZScoreTx}
\alias{plotShiftedZScoreTx}
\title{Plot shifted z scores}
\usage{
plotShiftedZScoreTx(shitedZScoresTx_results)
}
\arguments{
\item{shitedZScoresTx_results}{A \code{shitedZScoreTx.results} object.}
}
\value{
A plot.
}
\description{
Plot shifted z scores for permutation test results.
}
\examples{
file <- system.file(package="RgnTX", "extdata", "shiftedZScoreTx_results1.rds")
shiftedZScoreTx_results <- readRDS(file)
p1 <- plotShiftedZScoreTx(shiftedZScoreTx_results)
p1
}
\seealso{
\code{\link{shiftedZScoreTx}}
}
