% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftExonTx.R
\name{shiftExonTx}
\alias{shiftExonTx}
\title{Shift regions over transcripts}
\usage{
shiftExonTx(regions, start, width)
}
\arguments{
\item{regions}{A region set either to be \code{GRanges} or \code{GRangesList}, following the format indicated in vignette section 3.}

\item{start}{Starting positions. The function will return a region set shifting from these starting positions.}

\item{width}{Widths. The widths of returned regions. It should be positive/negative if users would like to move the starting positions to the right (3')/left (5').}
}
\value{
A \code{GRangesList} object.
}
\description{
Shift regions over transcripts (exons). This function accepts a feature set and outputs a region set from it. Each output region is from each input feature.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
exons.tx0 <- exonsBy(txdb)
#Shift regions over the following transcripts
trans.ids <- c("170", "1387", "4113", "10715")
regions <- exons.tx0[trans.ids]
start <- c(3624255, 55158197, 881641, 15694195)
width <- c(200, -200, 200, -200)
shifted_regions <- shiftExonTx(regions, start, width)
}
