% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2GRangesList.R
\name{vector2GRangesList}
\alias{vector2GRangesList}
\title{vector2GRangesList}
\usage{
vector2GRangesList(RefSeqID, targetName, strand, blockSizes, targetStart)
}
\arguments{
\item{RefSeqID}{The name of each element.}

\item{targetName}{The seqnames of each range.}

\item{strand}{The strand of each range.}

\item{blockSizes}{The width of each range.}

\item{targetStart}{The stard coordinate of each range.}
}
\value{
A \code{GRangesList} object.
}
\description{
Generate \code{GRangesList} object from vectors. The output region set follows the format required by the main permutation test functions.
}
\seealso{
\code{\link{GRanges2GRangesList}}
}
