% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateShift.R
\name{calculateShift}
\alias{calculateShift}
\title{Calculate positional shifting over transcriptome}
\usage{
calculateShift(regions, disp, direction = 'right', strand = '+')
}
\arguments{
\item{regions}{The region space where the displacement is calculated, which should be a GRangesList object.}

\item{disp}{The region space where the displacement is calculated, which should be a GRangesList object.}

\item{direction}{The direction of displacement. It has options 'left' and 'right'.}

\item{strand}{The strand type of the transcripts. It receives '+' or '-'.}
}
\value{
A \code{GRanges} object.
}
\description{
This function can calculate shifting over 'disconnected' regions. For example, due to alternative splicing, an mRNA is represented by several disconnected regions according to the annotation information. And this function can calculate position displacement over such disconnected regions. It can automatically skip the 'blank' between the regions when calculating position shifting.
}
\examples{
# Take five transcripts.
# Extract the last 200 nt regions from their CDS part.
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
trans.id.pstv <- c("170", "782", "974", "1364", "1387")
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene

# Download the CDS part of all transcriptome
cds.tx0 <- cdsBy(txdb, use.names = FALSE)

# pick the CDS part of these five transcripts
cds.p <- cds.tx0[trans.id.pstv]

width <- 200
disp.p.l <- data.frame(
    start = as.numeric(max(end(cds.p))),
    distance = width - 1,
    names = trans.id.pstv
)
R.p.l <- calculateShift(
    regions = cds.p, disp = disp.p.l,
    direction = "left", strand = "+"
)
}
\seealso{
\code{\link{extractRegions}}
}
