% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormatCorrect.R
\name{getFormatCorrect}
\alias{getFormatCorrect}
\title{getFormatCorrect}
\usage{
getFormatCorrect(A, B)
}
\arguments{
\item{A}{Region set 1. A Granges or GRangesList object.}

\item{B}{Region set 2. A Granges or GRangesList object.}
}
\value{
A \code{list} object.
}
\description{
This function makes sure the two input region sets are in the correct format required by RgnTX evaluation functions.
}
