% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getPvalZscore}
\alias{getPvalZscore}
\title{getPvalZscore}
\usage{
getPvalZscore(orig.ev, rand.ev, pval_t = FALSE)
}
\arguments{
\item{orig.ev}{Observed value.}

\item{rand.ev}{Random evaluation values.}

\item{pval_t}{Boolean. Default is FALSE. If FALSE, the p-value is calculated based on the number of random evaluations is larger or less than the initial evaluation. If TRUE, the p-value is calculated based on a t-test.}
}
\value{
A p-value and a z-score.
}
\description{
Calculate a p-value and z-score based on observed value and random evaluation values.
}
