% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPermResults.R
\name{plotPermResults}
\alias{plotPermResults}
\title{Plot permutation test results.}
\usage{
plotPermResults(permTestTx_results = NULL, breaks = 15, alpha = 0.05,
test_type = 'one-sided', binwidth = NULL)
}
\arguments{
\item{permTestTx_results}{A \code{permTestTx.results} list object, which can be generated by the permTestTx function.}

\item{breaks}{Histogram breaks. Default is 15.}

\item{alpha}{Significance level.}

\item{test_type}{The type of the test. This argument only receives either two options 'one-sided' or 'two-sided'. Default is 'one-sided'.}

\item{binwidth}{Histogram binwidth.}
}
\value{
A plot object.
}
\description{
Show a graphical representation of permutation test.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
exons.tx0 <- exonsBy(txdb)
trans.ids <- sample(names(exons.tx0), 500)
A <- randomizeTx(txdb, trans.ids, random_num = 100, random_length = 100)
B <- c(randomizeTx(txdb, trans.ids, random_num = 75, random_length = 100), A[1:25])
permTestTx_results <- permTestTx(A, B, txdb, ntimes = 10)
p1 <- plotPermResults(permTestTx_results, binwidth = 1)
}
\seealso{
\code{\link{permTestTx}}
}
