% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShiftedZScoreTx.R
\name{plotShiftedZScoreTx}
\alias{plotShiftedZScoreTx}
\title{Calculate shifted z scores.}
\usage{
plotShiftedZScoreTx(shitedZScoresTx_results)
}
\arguments{
\item{shitedZScoresTx_results}{A \code{shitedZScoreTx.results} object.}
}
\value{
A plot object.
}
\description{
Calculate shifted z scores for permutation test results.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
file <- system.file(package="RgnTX", "extdata/m6A_sites_data.rds")
m6A_sites_data <- readRDS(file)
RS1 <- m6A_sites_data[1:500]
permTestTx_results <- permTestTxIA_customPick(RS1 = RS1,
                                            txdb = txdb,
                                            customPick_function = getStopCodon,
                                            ntimes = 5)
shiftedZScoreTx_results <- shiftedZScoreTx(permTestTx_results, txdb = txdb,
                                            window = 2000,
                                            step = 200,
                                            ev_function_1 = overlapCountsTxIA)
p1 <- plotShiftedZScoreTx(shiftedZScoreTx_results)
}
\seealso{
\code{\link{shiftedZScoreTx}}
}
