% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2GRangesList.R
\name{vector2GRangesList}
\alias{vector2GRangesList}
\title{vector2GRangesList}
\usage{
vector2GRangesList(RefSeqID, targetName, strand, blockSizes, targetStart)
}
\arguments{
\item{RefSeqID}{The name of each element.}

\item{targetName}{The seqnames of each range.}

\item{strand}{The strand of each range.}

\item{blockSizes}{The width of each range.}

\item{targetStart}{The stard coordinate of each range.}
}
\value{
A \code{GRangesList} object.
}
\description{
Generate \code{GRangesList} object from vectors. The output region set follows the format required by the main permutation test functions.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
trans.tx0 <- transcriptsBy(txdb, "gene")[1:2]
trans.tx0.df <- data.frame(trans.tx0)
trans.tx0.id <- trans.tx0.df[, "tx_id"]
# RefSeqID
RefSeqID <- as.character(trans.tx0.id)
# targetName
targetName <- trans.tx0.df[, "seqnames"]
targetName <- as.character(targetName)
# strand
strand <- trans.tx0.df[, "strand"]
strand <- as.character(strand)
# blockSizes
blockSizes <- trans.tx0.df[, "width"]
# targetStart
targetStart <- trans.tx0.df[, "start"]

trans.list <- vector2GRangesList(RefSeqID, targetName, strand,
blockSizes, targetStart)
}
\seealso{
\code{\link{GRanges2GRangesList}}
}
